+++ modules/Students/includes/SaveEnrollment.fnc.php
@@ -22,6 +22,15 @@ function SaveEnrollment()
 				unset( $_REQUEST['month_values']['STUDENT_ENROLLMENT'][$stu_enrol_id] );
 				unset( $_REQUEST['year_values']['STUDENT_ENROLLMENT'][$stu_enrol_id] );
 			}
+			elseif ( $stu_enrol_id == 'new'
+				&& $_REQUEST['values']['STUDENT_ENROLLMENT']['new']['SCHOOL_ID'] )
+			{
+				if ( $_REQUEST['values']['STUDENT_ENROLLMENT']['new']['SCHOOL_ID'] !== UserSchool() )
+				{
+					// @since 5.4 Update current school to enrollment school.
+					$_SESSION['UserSchool'] = $_REQUEST['values']['STUDENT_ENROLLMENT']['new']['SCHOOL_ID'];
+				}
+			}
 			elseif ( UserStudentID() && ! empty( $stu_enrol_month['START_DATE'] ) )
 			{
 				$date = RequestedDate(